backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "dotenv>=0.9.9",
    "easyocr>=1.7.2",
    "fastapi>=0.118.0",
    "google-generativeai>=0.8.5",
    "pymupdf>=1.26.4",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.37.0",
]
